% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest Cross-Validation function}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{A numeric indicating the number of folds.}
}
\value{
A numeric indicating the mean square error.
}
\description{
This function performs a \code{k} fold random forest cross-validation
on the penguin data.
}
\examples{
my_rf_cv(2)
my_rf_cv(5)
my_rf_cv(10)

}
\keyword{prediction}
